refactor view namespaces
