+++ src/app/Library/Widget.php
@@ -3,7 +3,7 @@
 namespace Backpack\CRUD\app\Library;
 
 use Backpack\CRUD\app\Exceptions\BackpackProRequiredException;
-use Backpack\CRUD\app\Library\CrudPanel\ViewNamespaces;
+use Backpack\CRUD\ViewNamespaces;
 use Illuminate\Support\Fluent;
 
 /**
@@ -137,7 +137,7 @@ class Widget extends Fluent
         $type = $this->type;
         $paths = array_map(function ($item) use ($type) {
             return $item.'.'.$type;
-        }, ViewNamespaces::getViewNamespacesWithFallbackFor('widgets', 'backpack.base.component_view_namespaces.widgets'));
+        }, ViewNamespaces::getWithFallbackFor('widgets', 'backpack.base.component_view_namespaces.widgets'));
 
         foreach ($paths as $path) {
             if (view()->exists($path)) {
