+++ src/resources/views/crud/show.blade.php
@@ -61,7 +61,7 @@
                         		// including the configured view_namespaces
                         		$columnPaths = array_map(function($item) use ($column) {
                         			return $item.'.'.$column['type'];
-                        		}, \Backpack\CRUD\app\Library\CrudPanel\ViewNamespaces::getViewNamespacesFor('columns'));
+                        		}, \Backpack\CRUD\ViewNamespaces::getFor('columns'));
 
                         		// but always fall back to the stock 'text' column
                         		// if a view doesn't exist
